Extract collections/Read#index in a dedicated controller
It's more coherent with News and Bookmarks controllers.
